[Refactoring ⚙️] Machines 추천 페이지 컴포넌트 분리
