+++ src/components/Modal/components/Routine/RoutineAddModal.tsx
@@ -25,10 +25,13 @@ const RoutineAddModal = () => {
       setSelectedRoutines(new Set())
     },
   })
+
   const { onOpen } = useModal("루틴정보")
+
   const { setRoutineState, workoutState = { koreanName: "" } } = useModalStore()
+
   const { data: routines = [] } = useGetMyRoutines()
-  const { data: workouts, refetchAll } = useGetRoutineQueries(routines)
+  const { data: workouts } = useGetRoutineQueries(routines)
 
   const isFullRoutine = routines.length >= 5
 
@@ -58,10 +61,6 @@ const RoutineAddModal = () => {
     setRoutineState([...selectedRoutines])
     setSelectedRoutines(new Set())
   }
-
-  useEffect(() => {
-    refetchAll()
-  }, [routines, refetchAll])
 
   return (
     <Modal
