+++ src/hooks/mutation/usePostRecommend.ts
@@ -13,5 +13,11 @@ export const usePostRecommend = () => {
     onError: () => {
       console.error("에러")
     },
+    onSuccess: () => {
+      document.body.style.overflow = "auto"
+    },
+    onMutate: () => {
+      document.body.style.overflow = "hidden"
+    },
   })
 }
