+++ src/hooks/query/useGetMachineList.ts
@@ -7,7 +7,7 @@ import machineAPI from "@apis/domain/machine"
 export const useGetMachineList = () => {
   const { bodyPart } = useRecommendStore()
   return useQuery({
-    queryKey: ["getMachineList"],
+    queryKey: ["getMachineList", bodyPart],
     queryFn: async () =>
       await machineAPI.fetchList({ bodyPartKoreanName: bodyPart }),
     select: (data) => data.map((machine, idx) => ({ ...machine, id: idx })),
